Generate CHANGELOG file
I am preparing a new release but first wanted to make sure that we had
somesort of changelog history.

I have ended up using `github-changelog-generator` to generate it using
PRs, it isn't perfect as it doesn't include commits which haven't been
merged via a PR but it is better than nothing :-).